requirements.txt
streamlit
azure-ai-documentintelligence
pandas
openpyxl
xlrd
google-generativeai
python-docx